+++ CHANGELOG.md
@@ -10,3 +10,7 @@
 ## [0.0.1] - TODO: Add release date.
 
 * First commit.
+
+## [0.0.2] - Simplifying widget.
+
+* Simplifying widget.